--- actions/sonar/2.0/internal/sonar_issue.go
@@ -1,1 +1,0 @@
-package main
